requirements.txt
py==1.4.31
pysqlite==2.8.2
pytest==2.9.2
